billing and accounts app

+++ carts/models.py
@@ -61,7 +61,7 @@ m2m_changed.connect(m2m_changed_cart_receiver, sender=Cart.books.through)
 
 def pre_save_cart_receiver(sender, instance, *args, **kwargs):
     if instance.subtotal > 0:
-        instance.total = Decimal(instance.subtotal) * Decimal(1.10) # 10% tax
+        instance.total = float(instance.subtotal) * float(1.10) # 10% tax
     else:
         instance.total = 0.00
 pre_save.connect(pre_save_cart_receiver, sender=Cart)
